maybe it's the comment?
I wonder if it's the commit comment text that has the issue? This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work. This is a test to show how line breaks work.